virtual-data-analyst/requirements.txt
haystack-ai
python-dotenv
gradio
pandas
plotly
openpyxl
statsmodels
xlrd
psycopg2-binary
